+++ src/components/Skeleton/MovieList/MovieItem/SkeletonItem.ts
